Add files via upload
Project Title: Online Voting System.
The Online Voting System is a website or app that allows people to vote online instead of going to a voting booth. It is made to make voting easier, faster, and more secure. People can log in from anywhere, choose their favorite candidate, and submit their vote with just a few clicks.
Only registered users can vote, and the system makes sure each person can vote only once and see the results. Voters can see a list of candidates and get a message when their vote is successfully submitted.
Goal:
The goal of this project is to make voting simple, safe, and available to everyone online. It is useful for school elections, club voting, or even bigger elections with added security.